+++ Include/internal/pycore_interp.h
@@ -8,39 +8,54 @@ extern "C" {
 #  error "this header requires Py_BUILD_CORE define"
 #endif
 
-#include <stdbool.h>                  // bool
+#include <stdbool.h>              // bool
 
-#include "pycore_ast_state.h"         // struct ast_state
-#include "pycore_atexit.h"            // struct atexit_state
-#include "pycore_ceval_state.h"       // struct _ceval_state
-#include "pycore_code.h"              // struct callable_cache
-#include "pycore_context.h"           // struct _Py_context_state
-#include "pycore_crossinterp.h"       // struct _xidregistry
-#include "pycore_dtoa.h"              // struct _dtoa_state
-#include "pycore_exceptions.h"        // struct _Py_exc_state
-#include "pycore_floatobject.h"       // struct _Py_float_state
-#include "pycore_freelist.h"          // struct _Py_freelist_state
-#include "pycore_function.h"          // FUNC_MAX_WATCHERS
-#include "pycore_gc.h"                // struct _gc_runtime_state
-#include "pycore_genobject.h"         // struct _Py_async_gen_state
-#include "pycore_global_dict_state.h" // struct _Py_global_dict_state
-#include "pycore_global_objects.h"    // struct _Py_interp_cached_objects
-#include "pycore_import.h"            // struct _import_state
-#include "pycore_instruments.h"       // _PY_MONITORING_EVENTS
-#include "pycore_list.h"              // struct _Py_list_state
-#include "pycore_mimalloc.h"          // struct _mimalloc_interp_state
-#include "pycore_object_state.h"      // struct _py_object_state
-#include "pycore_obmalloc.h"          // struct _obmalloc_state
-#include "pycore_tstate.h"            // _PyThreadStateImpl
-#include "pycore_tuple.h"             // struct _Py_tuple_state
-#include "pycore_typeobject.h"        // struct types_state
-#include "pycore_unicodeobject.h"     // struct _Py_unicode_state
-#include "pycore_warnings.h"          // struct _warnings_runtime_state
+#include "pycore_ast_state.h"     // struct ast_state
+#include "pycore_atexit.h"        // struct atexit_state
+#include "pycore_ceval_state.h"   // struct _ceval_state
+#include "pycore_code.h"          // struct callable_cache
+#include "pycore_context.h"       // struct _Py_context_state
+#include "pycore_crossinterp.h"   // struct _xidregistry
+#include "pycore_dtoa.h"          // struct _dtoa_state
+#include "pycore_exceptions.h"    // struct _Py_exc_state
+#include "pycore_floatobject.h"   // struct _Py_float_state
+#include "pycore_freelist.h"      // struct _Py_freelist_state
+#include "pycore_function.h"      // FUNC_MAX_WATCHERS
+#include "pycore_gc.h"            // struct _gc_runtime_state
+#include "pycore_genobject.h"     // struct _Py_async_gen_state
+#include "pycore_global_objects.h"// struct _Py_interp_cached_objects
+#include "pycore_import.h"        // struct _import_state
+#include "pycore_instruments.h"   // _PY_MONITORING_EVENTS
+#include "pycore_list.h"          // struct _Py_list_state
+#include "pycore_mimalloc.h"      // struct _mimalloc_interp_state
+#include "pycore_object_state.h"  // struct _py_object_state
+#include "pycore_obmalloc.h"      // struct _obmalloc_state
+#include "pycore_tstate.h"        // _PyThreadStateImpl
+#include "pycore_tuple.h"         // struct _Py_tuple_state
+#include "pycore_typeobject.h"    // struct types_state
+#include "pycore_unicodeobject.h" // struct _Py_unicode_state
+#include "pycore_warnings.h"      // struct _warnings_runtime_state
 
 
 struct _Py_long_state {
     int max_str_digits;
 };
+
+#define DICT_MAX_WATCHERS 8
+
+struct _Py_dict_interp_state {
+    /*Global counter used to set ma_version_tag field of dictionary.
+     * It is incremented each time that a dictionary is created and each
+     * time that a dictionary is modified. */
+    uint64_t global_version;
+    uint32_t next_keys_version;
+    PyDict_WatchCallback watchers[DICT_MAX_WATCHERS];
+};
+
+#define _dict_state_INIT \
+    { \
+        .next_keys_version = 2, \
+    }
 
 
 /* cross-interpreter data registry */
